% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_wisedash_enrollment}
\alias{download_wisedash_enrollment}
\title{Download WISEdash enrollment data (2016+)}
\usage{
download_wisedash_enrollment(end_year)
}
\arguments{
\item{end_year}{School year end (2016+)}
}
\value{
Data frame with enrollment data
}
\description{
Downloads enrollment data from WISEdash data files. These are ZIP files
containing CSV data with enrollment by grade level.
}
\keyword{internal}
