% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_published_enr}
\alias{process_published_enr}
\title{Process published enrollment data (1997-2015)}
\usage{
process_published_enr(df, end_year)
}
\arguments{
\item{df}{Raw Excel data frame}

\item{end_year}{School year end}
}
\value{
Processed data frame
}
\description{
Published Excel files are in wide format with grades as columns.
}
\keyword{internal}
